build(deps): rusqlite 0.40 and a single reqwest 0.13
- rusqlite 0.31 → 0.40 moves the bundled SQLite from 3.45.0 to 3.53.2,
  past CVE-2025-29087 and CVE-2025-6965. The only API change that
  reached the code is that usize no longer converts from SQL, in tests.
- reqwest 0.12 → 0.13 removes the second copy the updater already
  pulled in. Like the updater, it is built without a crypto provider:
  http_client installs ring once and every client in the app, tests and
  examples comes from there. Certificates are now checked against the
  system store instead of bundled roots. HTTPS to GitHub and Hugging Face
  was checked with a real request.

+++ desktop/src-tauri/src/db.rs
@@ -713,7 +713,7 @@ mod tests {
 
         let error = migrate_from_json(&conn, tmp.path()).unwrap_err();
         assert!(error.contains("insert history[2]"));
-        let count: usize = conn
+        let count: i64 = conn
             .query_row("SELECT COUNT(*) FROM history", [], |r| r.get(0))
             .unwrap();
         assert_eq!(
@@ -725,13 +725,13 @@ mod tests {
 
         conn.execute_batch("DROP TRIGGER reject_second").unwrap();
         migrate_from_json(&conn, tmp.path()).unwrap();
-        let count: usize = conn
+        let count: i64 = conn
             .query_row("SELECT COUNT(*) FROM history", [], |r| r.get(0))
             .unwrap();
         assert_eq!(count, 2);
         assert!(!path.exists());
         migrate_from_json(&conn, tmp.path()).unwrap();
-        let count: usize = conn
+        let count: i64 = conn
             .query_row("SELECT COUNT(*) FROM history", [], |r| r.get(0))
             .unwrap();
         assert_eq!(count, 2);
